feat(usecase): added is author check to DeleteCategoryUseCase

+++ src/main/kotlin/com/bortnik/todo/usecase/category/DeleteCategoryUseCase.kt
@@ -1,5 +1,7 @@
 package com.bortnik.todo.usecase.category
 
+import com.bortnik.todo.domain.exceptions.category.CategoryNotFound
+import com.bortnik.todo.domain.exceptions.user.AcceptError
 import com.bortnik.todo.domain.repositories.CategoryRepository
 import org.springframework.stereotype.Service
 
@@ -8,7 +10,12 @@ class DeleteCategoryUseCase(
     private val categoryRepository: CategoryRepository
 ) {
 
-    fun deleteCategory(categoryId: Int) {
+    fun deleteCategory(categoryId: Int, userId: Int) {
+        val category = categoryRepository.getCategoryById(categoryId)
+            ?: throw CategoryNotFound("category to delete with id '$categoryId' not found")
+        if (category.userId != userId) {
+            throw AcceptError("You dont have access rights to delete this category")
+        }
         categoryRepository.deleteCategory(categoryId)
     }
